+++ src/components/layout/Header.js
@@ -5,6 +5,7 @@ import logoMT from '../../resources/mytill-logo-transparent.png';
 import { useNavigate } from "react-router";
 import { COLOR_PALETTE } from "../../Constants";
 import React, { useState } from "react";
+import { saveUser, login } from "../../requests/users-req";
 
 const useStyles = makeStyles({
     toolBar: {
@@ -59,6 +60,25 @@ const Header = () => {
     
     const handleSignUp = () => {
         navigate('/create-account');
+    }
+
+    // TODO
+    async function handleTemp () {
+        let res = await saveUser({
+            fname: 'test',
+            lname: 'icle',
+            email: 'testicle@domain.cum'+Math.floor(Math.random()*100),
+            password: 'temp123'
+        })
+        console.log(res);
+    }
+
+    async function handleTemp2 () {
+        let res = await login({
+            email: 'testicle@domain.cum',
+            password: 'temp123'
+        })
+        console.log(res);
     }
 
     const handleHome = () => {