+++ src/CHANGELOG.md
@@ -1,7 +1,12 @@
  # CHANGELOG
 This changelog will attempt to use [Semantic Versioning](https://semver.org/). This will help us keep track of what has been completed up to the date shown at the top of the list of tickets. 🖥️
 
+Create a new version whenever we update main/master.
+
 Add items closer to the version number so the date go from newest to oldest (top to bottom).
 
-### *0.1.0*
+### *Version 0.1.1*
+  - 2022/11/28 - [Create Reuseable components](https://sd-onlinepos.atlassian.net/browse/ON-28) 
+  - 2022/11/28 - [Create Layout for Landing Page](https://sd-onlinepos.atlassian.net/browse/ON-41) 
+  - 2022/11/20 - [Create Layout for Create Account Page](https://sd-onlinepos.atlassian.net/browse/ON-45) 
   - 2022/11/20 - [Create CHANGELOG.md file](https://sd-onlinepos.atlassian.net/browse/ON-40)